Match Rocket Belt Files (#144)
* start rocket belt

* rocket belt pass

* post merge fix

* cleanup some fakematches

* add prototypes

* format

* Match func_8032975C

* misc cleanup

+++ src/app/hover_pads.h
@@ -35,7 +35,7 @@ void hoverPadFrameUpdate(Mtx4F*);
 void hoverPadLanded(s32 hoverIdx);
 f32 hoverPadGetAltitude(s32 hoverIdx);
 f32 hoverPadGetFuel(s32 hoverIdx);
-s32 hoverPadGetLandedIdx(s32 objId, s32);
+s32 hoverPadGetLandedIdx(s32 objId, f32);
 void hoverPadDeinit(void);
 u8 hoverPadGetCount(void);
 s16 hoverPadGetPoints(void);
